requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
numpy==1.19.2
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.44
Pillow==7.2.0
pymongo==3.11.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
vk-api==11.9.0

scikit-learn~=0.23.2